requirements_STT.txt
torch==2.6.0
transformers==4.51.3
librosa==0.10.0
numpy==1.26.4
pyzmq==26.2.1
opencc==1.1.9
soundfile==0.13.1
audioread==3.0.1
safetensors==0.5.3
accelerate==1.5.2
tqdm==4.67.1